Move screen/actions to neovim folder
