+++ browser/src/Editor/NeovimInput.tsx
@@ -7,8 +7,7 @@
 import * as React from "react"
 
 import { Mouse } from "./../Input/Mouse"
-import { NeovimInstance } from "./../neovim"
-import { NeovimScreen } from "./../Screen"
+import { NeovimInstance, NeovimScreen } from "./../neovim"
 
 import { TypingPredictionManager } from "./../Services/TypingPredictionManager"
