+++ browser/src/neovim/index.ts
@@ -1,5 +1,6 @@
 export * from "./EventContext"
 export * from "./MsgPack"
+export * from "./Screen"
 export * from "./Session"
 export * from "./NeovimProcessSpawner"
 export * from "./NeovimInstance"
